[BE/FIX/#12] ENTITY - ERD 일치

+++ Backend/src/main/java/com/hexa/muinus/common/exception/StoreLocationDuplicateException.java
@@ -2,8 +2,10 @@ package com.hexa.muinus.common.exception;
 
 import org.springframework.http.HttpStatus;
 
+import java.math.BigDecimal;
+
 public class StoreLocationDuplicateException extends MuinusException {
-    public StoreLocationDuplicateException(double x, double y) {
+    public StoreLocationDuplicateException(BigDecimal x, BigDecimal y) {
         super("이미 등록된 위치입니다. locationX:" + x + ", locationY:" + y, 400);
     }
 }
