+++ Backend/src/main/java/com/hexa/muinus/store/dto/StoreSearchDTO.java
@@ -6,14 +6,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class StoreSearchDTO {
     private Integer storeNo;
     private String storeName;
-    private Double locationX;
-    private Double locationY;
+    private BigDecimal locationX;
+    private BigDecimal locationY;
     private String address;
     private String phone;
     private String itemName;
